pyproject.toml
[tool.poetry]
name = "repo-review"
version = "0.1.0"
description = ""
authors = ["Foxicution <foxicution@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
networkx = "^2.8.8"
openai = "^0.25.0"
protobuf = "3.20.*"
PyGithub = "^1.57"
pyvis = "^0.3.1"
streamlit = "^1.15.2"
toml = "^0.10.2"
toolz = "^0.12.0"
google-cloud-firestore = "^2.7.3"
lambdas = "^0.1.0"
option = "^2.1.0"
fastapi = {extras = ["all"], version = "^0.88.0"}
jinja2 = "^3.1.2"
tree-sitter-languages = "^1.5.0"
tree-sitter = "^0.20.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
ipykernel = "^6.19.2"
pre-commit = "^2.20.0"
scalene = "^1.5.16"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 94
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
